Create vm_getmacinfo.yml
This is the task file for nicinfo.yml, output for nicinfo.yml will be created in /tmp/tempouts/ directory for each VM.
This playbook uses both vm_guest_facts ansible module as well as VMware Vcenter RestApi. 
This will require appropriate permissions to your account in Vcenter as pre-requisite.
This playbook is not modifying or changing anything on Vcenter config. This playbook is purely to fetch the NIC and DVS_PORT/VLAN mappings from VMware Vcenter for a Virtual Machine. 
Minimum supported Vcenter is 6.5, as RestAPI support not available prior to that version.