VMHOSTinfo.yml
This playbook will fetch esxi hosts of multiple VMs provided in inputfile and will create outputfile as specified in parameters.
Actually this is a sample for fetching vm_guest facts for multiple VMs and getting them in a sort of tabular format. getexsihost.yml is the taskfile for this playbook.